Move internal/inventory to frontend
Only used there, so moving it.

Test plan:

Just moved code, CI and compiler will find issues.
