+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -239,6 +239,7 @@ go_library(
         "//cmd/frontend/hubspot",
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/cloneurls",
+        "//cmd/frontend/internal/inventory",
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
         "//internal/actor",
@@ -295,7 +296,6 @@ go_library(
         "//internal/honey/search",
         "//internal/httpcli",
         "//internal/insights",
-        "//internal/inventory",
         "//internal/jsonc",
         "//internal/lazyregexp",
         "//internal/licensing",
@@ -478,6 +478,7 @@ go_test(
         "//cmd/frontend/graphqlbackend/externallink",
         "//cmd/frontend/graphqlbackend/graphqlutil",
         "//cmd/frontend/internal/guardrails",
+        "//cmd/frontend/internal/inventory",
         "//internal/actor",
         "//internal/api",
         "//internal/auth",
@@ -507,7 +508,6 @@ go_test(
         "//internal/gitserver/protocol",
         "//internal/gqlutil",
         "//internal/highlight",
-        "//internal/inventory",
         "//internal/observation",
         "//internal/oobmigration",
         "//internal/ratelimit",
