+++ cmd/frontend/internal/inventory/BUILD.bazel
@@ -8,8 +8,8 @@ go_library(
         "entries.go",
         "inventory.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/inventory",
-    visibility = ["//:__subpackages__"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/inventory",
+    visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//lib/errors",
         "@com_github_go_enry_go_enry_v2//:go-enry",
@@ -20,7 +20,6 @@ go_library(
 
 go_test(
     name = "inventory_test",
-    timeout = "short",
     srcs = [
         "entries_test.go",
         "inventory_test.go",
