+++ cmd/frontend/internal/inventory/context.go
