+++ cmd/frontend/internal/inventory/entries.go
