+++ cmd/frontend/internal/inventory/entries_test.go
