+++ cmd/frontend/internal/inventory/inventory.go
